CPU hours tracking

+++ db/events.go
@@ -18,7 +18,7 @@ type CPUUsageEvent struct {
 	RecordDate    time.Time `db:"record_date" json:"record_date"`
 	EffectiveDate time.Time `db:"effective_date" json:"effective_date"`
 	EventType     EventType `db:"event_type" json:"event_type"`
-	Value         int64     `db:"value" json:"value"`
+	Value         float64   `db:"value" json:"value"`
 	CreatedBy     string    `db:"created_by" json:"created_by"`
 	LastModified  string    `db:"last_modified" json:"last_modified"`
 }
